Extract LED handling from ESS main class
We originally added some LED handling the the ESS main class (SensirionESS)
with the goal to simplify interfacing with the onboard LEDs on the ESS.
While that seemed to be a good idea at the time, it really only works when
the ESS is used on a board that follows the Arduino Uno footprint. Many
other common Arduino-compatible boards have different pinouts, and some
of the LED pin settings can conflict with I2C ports (e.g. on ESP8266 or
MKR board).

This commit moves this into a separate class (SensirionEssLeds), renames
the constant for the Arduino Uno pins to include ARDUINO_UNO_ in their
name, and requires passing those pins in the constructor to force the
user to think about it while setting it up. This is a bit more code to
write, but since we provide a fairly complete example in 'led-demo' I
believe people can just copy this anyway.